Ditch fixture.py
